separate pretoks testing into different test file
