+++ tests/test_iterdictindex.py
@@ -186,102 +186,6 @@ class TestIterDictIndexer(TempDirTestCase):
         #restore setting after test
         pt.ApplicationSetup.setProperty("termpipelines", "Stopwords,PorterStemmer")
         self.assertEqual("", index.getIndexProperty("termpipelines", "bla"))
-        
-
-    def test_json_pretok_iterator(self):
-        it = [
-            {'docno': 'd1', 'url': 'url1', "toks" : {"a" : 1, "b" : 2}}
-        ]
-
-        import json
-        from pyterrier import autoclass
-        testfile = os.path.join(self.test_dir, "test.json")
-        with open(testfile, 'wt') as file:
-            for doc in it:
-                file.write(json.dumps(doc))
-        jparserCls = autoclass("org.terrier.python.JsonlPretokenisedIterator")
-        jparser = jparserCls(testfile)
-        self.assertTrue(jparser.hasNext())
-        nextRow = jparser.next()
-        self.assertIsNotNone(nextRow)
-        props = nextRow.getKey()
-        self.assertEqual("d1", props.get("docno"))
-        self.assertEqual("url1", props.get("url"))
-        pl = nextRow.getValue()
-        self.assertIn("a", pl.termSet())
-        self.assertEqual(1, pl.getFrequency("a"))
-
-    def _make_pretok_index(self, n, index_type, meta=('docno', 'url')):
-        from pyterrier.index import IndexingType
-        # Test both versions: _fifo (for UNIX) and _nofifo (for Windows)
-        indexers = [
-            pt.index._IterDictIndexer_fifo(self.test_dir, type=index_type, meta=meta, pretokenised=True),
-            pt.index._IterDictIndexer_fifo(self.test_dir, type=index_type, threads=4, meta=meta, pretokenised=True),
-            pt.index._IterDictIndexer_nofifo(self.test_dir, type=index_type, meta=meta, pretokenised=True),
-        ]
-        if BaseTestCase.is_windows():
-           indexers = [indexers[-1]] 
-        for indexer in indexers:
-            with self.subTest(indexer=indexer):
-                it = [
-                    {'docno': 'd1', 'url': 'url1', "toks" : {"a" : 1, "b" : 2}}
-                ]
-                it = itertools.islice(it, n)
-                print("Writing index to " + self.test_dir)
-                indexref = indexer.index(it)
-                self.assertIsNotNone(indexref)
-                index = pt.IndexFactory.of(indexref)
-                self.assertIsNotNone(index)
-                self.assertEqual(n, index.getCollectionStatistics().getNumberOfDocuments())
-                self.assertEqual(len(set(meta)), len(index.getMetaIndex().getKeys()))
-                for m in meta:
-                    self.assertTrue(m in index.getMetaIndex().getKeys())
-                if index_type is IndexingType.CLASSIC:
-                    self.assertTrue(os.path.isfile(self.test_dir + '/data.direct.bf'))
-                    self.assertTrue(index.hasIndexStructure("direct"))
-                else:
-                    self.assertFalse(os.path.isfile(self.test_dir + '/data.direct.bf'))
-                inv = index.getInvertedIndex()
-                lex = index.getLexicon()
-                post = inv.getPostings(lex.getLexiconEntry('a'))
-                post.next()
-                self.assertEqual(1, post.frequency)
-                index.close()
-            # reset index directory for next run
-            shutil.rmtree(self.test_dir)
-            os.mkdir(self.test_dir)
-
-    def test_pretok_createindex1(self):
-        from pyterrier.index import IndexingType
-        from jnius import JavaException
-        try:
-            self._make_pretok_index(1, IndexingType.CLASSIC)
-        except JavaException as ja:
-            print("\n\t".join(ja.stacktrace))
-
-    def test_pretok_createindex2(self):
-        from pyterrier.index import IndexingType
-        self._make_pretok_index(2, IndexingType.CLASSIC)
-
-    def test_pretok_createindex3(self):
-        from pyterrier.index import IndexingType
-        self._make_pretok_index(3, IndexingType.CLASSIC)
-
-    def test_pretok_createindex1_single_pass(self):
-        from pyterrier.index import IndexingType
-        from jnius import JavaException
-        try:
-            self._make_pretok_index(1, IndexingType.SINGLEPASS)
-        except JavaException as ja:
-            print("\n\t".join(ja.stacktrace))
-
-    def test_pretok_createindex2_single_pass(self):
-        from pyterrier.index import IndexingType
-        self._make_pretok_index(2, IndexingType.SINGLEPASS)
-
-    def test_pretok_createindex3_single_pass(self):
-        from pyterrier.index import IndexingType
-        self._make_pretok_index(3, IndexingType.SINGLEPASS)
 
 if __name__ == "__main__":
     unittest.main()
